requirements.txt
# Core GenAI Libraries
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Additional AI Libraries
transformers>=4.35.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Web Framework (for API development)
flask>=3.0.0
requests>=2.31.0

ibm-watsonx-ai==1.0.8
ibm-watson-machine-learning==1.0.367
langchain-ibm==0.1.7
langchain-community==0.2.10
langchain-experimental==0.0.62
langchainhub==0.1.18
langchain==0.2.11
pypdf==4.2.0
chromadb==0.4.24
tenacity==8.2.3
langchain-core==0.2.43
# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0

# Jupyter for notebooks
jupyter>=1.0.0
notebook>=7.0.0

gradio==4.44.0
pydantic==2.9.1
huggingface_hub==0.23.0
